model/diffusion.py
import torch
import torch.nn as nn

class DiscreteDiffusion(nn.Module):
    def __init__(self, num_steps, vocab_size):
        super().__init__()
        self.num_steps = num_steps
        self.vocab_size = vocab_size
        
        # Define transition matrices
        self.register_buffer('Q', self._get_transition_matrix())
        
    def _get_transition_matrix(self):
        # Create transition matrix for token corruption
        # This is a simplified version - you might want to use a more sophisticated approach
        Q = torch.full((self.vocab_size, self.vocab_size), 1/self.vocab_size)
        Q = Q + torch.eye(self.vocab_size) * 0.5
        Q = Q / Q.sum(dim=1, keepdim=True)
        return Q
    
    def q_sample(self, x_0, t):
        """Corrupt input tokens to timestep t"""
        # Implementation of forward diffusion process
        pass
    
    def p_sample(self, model, x_t, t, protein_embedding, interface_mask):
        """Single step of reverse diffusion"""
        # Implementation of reverse diffusion step
        pass