requirements.txt
requests>=2.31.0
tqdm>=4.66.0
arxiv>=1.4.8
biopython>=1.82.0
rdkit>=2024.03.1
torch_geometric
esm
wandb
transformers